저금통 추가
bank.lua
coin.lua -modify
springAlgorithm - add
