+++ Coin.lua
@@ -62,42 +62,3 @@ end
 function Coin:Update()
   self.y = self.y + 1 -- 계속 떨어져야 하므로.
 end
-
-function SetCoinAlgorithmDefault()
-	MaxCoin = love.math.random(5,10) -- 떨어지는 코인의 개수
-	KeepCoin = 0
-	coinCount = 0
-	print("MaxCoin : "..MaxCoin)
-	print("KeepCoin :"..KeepCoin)
-	print("coinCount : "..coinCount)
-end
-
-function SpringExplainAlgorithm()
-	love.graphics.setColor(255, 0, 0, 255)
-	love.graphics.print("하늘에서 동전이 저금통으로 들어가요! \n다 떨어지고 난 후에 저금통에 얼마가 있을까요?",62+285, 13)
-	love.graphics.print("현재까지 모은 돈 : "..KeepCoin,62+285,50)
-	love.graphics.print("현재까지 떨어진 동전 : "..coinCount,62+285,100)
-end
-
-function StageSpringAlgorithm()
-	if stageLevel ~= 5 then
-		return
-	end
-	SpringExplainAlgorithm()
-
-	if coin == nil and coinCount < MaxCoin then
-			coin = Coin.Create()
-			KeepCoin = KeepCoin + coin:GetPoint()
-			coinCount = coinCount + 1
-	elseif coin ~= nil then
-		coin:Draw()
-		coin:Update()
-
-		if coin:GetY() == 130 then
-					coin = nil
-		end
-	elseif coin == nil and coinCount == MaxCoin then
-			-- PrintAnswer()
-			-- 정답 선택을 어떻게 할것인가?
-	end
-end
